requirements.txt
Flask>=2.3
discord.py>=2.4
python-dotenv>=1.0
markdown>=3.5
pymdown-extensions>=9.3
Werkzeug>=2.3
Jinja2>=3.1
MarkupSafe>=2.1
click>=8.1
itsdangerous>=2.1
colorama>=0.4.6
aiohttp>=3.8
frozenlist>=1.4
multidict>=6.0
yarl>=1.8
attrs>=23.1
charset-normalizer>=3.0
idna>=3.4
requests>=2.31
urllib3>=2.0
certifi>=2023.7